add bundlePath to GetBundleForChannel response
When we deprecated the GetBundleForChannel GRPC endpoint, we also pruned
a significant amount of data from the sqlite GRPC server response in order
to improve on-cluster performance. Unfortunately we've become aware that
users depending on the response containing the bundle path field.

This commit:
1. adds the bundle path field back to that response.
2. updates the FBC GRPC endpoint to exhibit the same behavior
3. updates the server tests to ensure the bundle path is present in the
   response for both FBC and sqlite implementations

While working on item (3), the required change ended up breaking the
sqlite ListBundles test. It turns out that our server tests were not
using the same configuration as the `opm registry serve` command. In
`opm registry serve` we omit the csv and object fields from the response
when the bundle includes a bundle path. This commit updates the test
configuration to _also_ omit bundles when a bundle path is present.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>